hoprd_migration::sea_orm
Module entity

Module for the Entity type and operations

Modules

prelude

Re-export common types from the entity

Structs

ColumnDef

Defines a Column for an Entity

RelationBuilder

Defines a helper to build a relation

RelationDef

Defines a relationship

Enums

ActiveValue

Defines a stateful value used in ActiveModel.

ColumnType

All column types

Identity

List of column identifier

RelationType

Defines the type of relationship

Value

Value variants

Traits

ActiveEnum

A Rust representation of enum defined in database.

ActiveEnumValue

The Rust Value backing ActiveEnums

ActiveModelBehavior

A Trait for overriding the ActiveModel behavior

ActiveModelTrait

A Trait for ActiveModel to perform Create, Update or Delete operation. The type must also implement the EntityTrait. See module level docs crate::entity for a full example

ColumnTrait

API for working with a Column. Mostly a wrapper of the identically named methods in sea_query::Expr

ColumnTypeTrait

SeaORM’s utility methods that act on ColumnType

EntityName

A Trait for mapping an Entity to a database table

EntityTrait

An abstract base class for defining Entities.

FromQueryResult

A Trait for implementing a QueryResult

IdenStatic

Ensure the identifier for an Entity can be converted to a static str

IdentityOf

Check the Identity of an Entity

IntoActiveModel

A Trait for any type that can be converted into an ActiveModel

IntoActiveValue

Any type that can be converted into an ActiveValue

IntoIdentity

Performs a conversion into an Identity

Iterable

This trait designates that an Enum can be iterated over. It can be auto generated using the EnumIter derive macro.

Linked

A Trait for links between Entities

ModelTrait

A Trait for a Model

PartialModelTrait

A trait for a part of Model

PrimaryKeyArity

How many columns this Primary Key comprises

PrimaryKeyToColumn

How to map a Primary Key to a column

PrimaryKeyTrait

A Trait for to be used to define a Primary Key.

Related

Checks if Entities are related

RelationTrait

Defines the relations of an Entity

TryIntoModel

A Trait for any type that can be converted into an Model

Functions

Set

Defines a set operation on an ActiveValue

Unchanged

Defines an unchanged operation on an ActiveValue

UnsetDeprecated

Defines an not set operation on an ActiveValue

Type Aliases

ForeignKeyAction

Action to perform on a foreign key whenever there are changes to an ActiveModel

LinkDef

Same as RelationDef