hoprd_migration::sea_orm::entity
Struct RelationBuilder

pub struct RelationBuilder<E, R>
where
    E: EntityTrait,
    R: EntityTrait,
{ /* private fields */ }

Defines a helper to build a relation

Implementations

impl<E, R> RelationBuilder<E, R>
where E: EntityTrait, R: EntityTrait,

pub fn from<T>(self, identifier: T) -> RelationBuilder<E, R>
where T: IdentityOf<E>,

Build a relationship from an Entity

pub fn to<T>(self, identifier: T) -> RelationBuilder<E, R>
where T: IdentityOf<R>,

Build a relationship to an Entity

pub fn on_delete(self, action: ForeignKeyAction) -> RelationBuilder<E, R>

An operation to perform on a foreign key when a delete operation occurs

pub fn on_update(self, action: ForeignKeyAction) -> RelationBuilder<E, R>

An operation to perform on a foreign key when an update operation occurs

pub fn on_condition<F>(self, f: F) -> RelationBuilder<E, R>
where F: Fn(SeaRc<dyn Iden>, SeaRc<dyn Iden>) -> Condition + 'static + Send + Sync,

Set custom join ON condition.

This method takes a closure with parameters denoting the left-hand side and right-hand side table in the join expression.

pub fn fk_name(self, fk_name: &str) -> RelationBuilder<E, R>

Set the name of foreign key constraint

pub fn condition_type( self, condition_type: ConditionType, ) -> RelationBuilder<E, R>

Set the condition type of join on expression