hoprd_migration::sea_orm::entity
Trait EntityName

pub trait EntityName: IdenStatic + Default {
    // Required method
    fn table_name(&self) -> &str;

    // Provided methods
    fn schema_name(&self) -> Option<&str> { ... }
    fn comment(&self) -> Option<&str> { ... }
    fn module_name(&self) -> &str { ... }
    fn table_ref(&self) -> TableRef { ... }
}

A Trait for mapping an Entity to a database table

Required Methods

fn table_name(&self) -> &str

Get the name of the table

Provided Methods

fn schema_name(&self) -> Option<&str>

Method to get the name for the schema, defaults to Option::None if not set

fn comment(&self) -> Option<&str>

Method to get the comment for the schema, defaults to Option::None if not set

fn module_name(&self) -> &str

Get the name of the module from the invoking self.table_name()

fn table_ref(&self) -> TableRef

Get the TableRef from invoking the self.schema_name()

Object Safety

This trait is not object safe.