hoprd_migration::sea_orm
Enum DbErr

pub enum DbErr {

    ConnectionAcquire(ConnAcquireErr),
    TryIntoErr {
        from: &'static str,
        into: &'static str,
        source: Box<dyn Error + Sync + Send>,
    },
    Conn(RuntimeErr),
    Exec(RuntimeErr),
    Query(RuntimeErr),
    ConvertFromU64(&'static str),
    UnpackInsertId,
    UpdateGetPrimaryKey,
    RecordNotFound(String),
    AttrNotSet(String),
    Custom(String),
    Type(String),
    Json(String),
    Migration(String),
    RecordNotInserted,
    RecordNotUpdated,

}

An error from unsuccessful database operations

Variants

ConnectionAcquire(ConnAcquireErr)

This error can happen when the connection pool is fully-utilized

TryIntoErr

Runtime type conversion error

Fields

from: &'static str

From type

into: &'static str

Into type

source: Box<dyn Error + Sync + Send>

TryError

Conn(RuntimeErr)

There was a problem with the database connection

Exec(RuntimeErr)

An operation did not execute successfully

Query(RuntimeErr)

An error occurred while performing a query

ConvertFromU64(&'static str)

Type error: the specified type cannot be converted from u64. This is not a runtime error.

UnpackInsertId

After an insert statement it was impossible to retrieve the last_insert_id

UpdateGetPrimaryKey

When updating, a model should know its primary key to check if the record has been correctly updated, otherwise this error will occur

RecordNotFound(String)

The record was not found in the database

AttrNotSet(String)

Thrown by TryFrom<ActiveModel>, which assumes all attributes are set/unchanged

Custom(String)

A custom error

Type(String)

Error occurred while parsing value as target type

Json(String)

Error occurred while parsing json value as target type

Migration(String)

A migration error

RecordNotInserted

None of the records are inserted, that probably means all of them conflict with existing records in the table

RecordNotUpdated

None of the records are updated, that means a WHERE condition has no matches. May be the table is empty or the record does not exist

Implementations

impl DbErr

pub fn sql_err(&self) -> Option<SqlErr>

Convert generic DbErr by sqlx to SqlErr, return none if the error is not any type of SqlErr