hoprd_migration::sea_orm::schema
Struct Schema

pub struct Schema { /* private fields */ }

This is a helper struct to convert EntityTrait into different sea_query statements.

Implementations

impl Schema

pub fn create_enum_from_active_enum<A>(&self) -> TypeCreateStatement
where A: ActiveEnum,

Creates Postgres enums from an ActiveEnum. See TypeCreateStatement for more details

pub fn create_enum_from_entity<E>(&self, entity: E) -> Vec<TypeCreateStatement>
where E: EntityTrait,

Creates Postgres enums from an Entity. See TypeCreateStatement for more details

pub fn create_table_from_entity<E>(&self, entity: E) -> TableCreateStatement
where E: EntityTrait,

Creates a table from an Entity. See TableCreateStatement for more details.

pub fn create_index_from_entity<E>( &self, entity: E, ) -> Vec<IndexCreateStatement>
where E: EntityTrait,

Creates the indexes from an Entity, returning an empty Vec if there are none to create. See IndexCreateStatement for more details

pub fn get_column_def<E>(&self, column: <E as EntityTrait>::Column) -> ColumnDef
where E: EntityTrait,

Creates a column definition for example to update a table.

use crate::sea_orm::IdenStatic;
use sea_orm::{
    ActiveModelBehavior, ColumnDef, ColumnTrait, ColumnType, DbBackend, EntityName,
    EntityTrait, EnumIter, PrimaryKeyTrait, RelationDef, RelationTrait, Schema,
};
use sea_orm_macros::{DeriveEntityModel, DerivePrimaryKey};
use sea_query::{MysqlQueryBuilder, TableAlterStatement};

#[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
#[sea_orm(table_name = "posts")]
pub struct Model {
    #[sea_orm(primary_key)]
    pub id: u32,
    pub title: String,
}

#[derive(Copy, Clone, Debug, EnumIter)]
pub enum Relation {}

impl RelationTrait for Relation {
    fn def(&self) -> RelationDef {
        panic!("No RelationDef")
    }
}
impl ActiveModelBehavior for ActiveModel {}

let schema = Schema::new(DbBackend::MySql);

let mut alter_table = TableAlterStatement::new()
    .table(Entity)
    .add_column(&mut schema.get_column_def::<Entity>(Column::Title))
    .take();
assert_eq!(
    alter_table.to_string(MysqlQueryBuilder::default()),
    "ALTER TABLE `posts` ADD COLUMN `title` varchar(255) NOT NULL"
);

impl Schema

pub fn new(backend: DatabaseBackend) -> Schema

Create a helper for a specific database backend