hoprd_migration::sea_orm::sea_query

Module foreign_key

Foreign key definition & alternations statements.

Usage

Table Foreign Key Create, see ForeignKeyCreateStatement

Table Foreign Key Drop, see ForeignKeyDropStatement

Structs

ForeignKey

Shorthand for constructing any foreign key statement

ForeignKeyCreateStatement

Create a foreign key constraint for an existing table. Unsupported by Sqlite

ForeignKeyDropStatement

Drop a foreign key constraint for an existing table

TableForeignKey

Specification of a foreign key

Enums

ForeignKeyAction

Foreign key on update & on delete actions

ForeignKeyStatement

All available types of foreign key statement