hoprd_migration::sea_orm
Crate sqlx

The async SQL toolkit for Rust, built with ❤️ by the LaunchBadge team.

See our README to get started or browse our example projects.
Have a question? Check our FAQ or open a discussion.

Runtime Support

SQLx supports both the Tokio and async-std runtimes.

You choose which runtime SQLx uses by default by enabling one of the following features:

runtime-async-std

runtime-tokio

The runtime-actix feature also exists but is an alias of runtime-tokio.

If more than one runtime feature is enabled, the Tokio runtime is used if a Tokio context exists on the current thread, i.e. tokio::runtime::Handle::try_current() returns Ok; async-std is used otherwise.

Note that while SQLx no longer produces a compile error if zero or multiple runtime features are enabled, which is useful for libraries building on top of it, the use of nearly any async function in the API will panic without at least one runtime feature enabled.

The chief exception is the SQLite driver, which is runtime-agnostic, including its integration with the query macros. However, SqlitePool does require runtime support for timeouts and spawning internal management tasks.

TLS Support

For securely communicating with SQL servers over an untrusted network connection such as the internet, you can enable Transport Layer Security (TLS) by enabling one of the following features:

tls-native-tls: Enables the native-tls backend which uses the OS-native TLS capabilities:

SecureTransport on macOS.

SChannel on Windows.

OpenSSL on all other platforms.

tls-rustls: Enables the RusTLS backend, a crossplatform TLS library.

Only supports TLS revisions 1.2 and 1.3.

If you get HandshakeFailure errors when using this feature, it likely means your database server does not support these newer revisions. This might be resolved by enabling or switching to the tls-native-tls feature.

If more than one TLS feature is enabled, the tls-native-tls feature takes precedent so that it is only necessary to enable it to see if it resolves the HandshakeFailure error without disabling tls-rustls.

Consult the user manual for your database to find the TLS versions it supports.

If your connection configuration requires a TLS upgrade but TLS support was not enabled, the connection attempt will return an error.

The legacy runtime+TLS combination feature flags are still supported, but for forward-compatibility, use of the separate runtime and TLS feature flags is recommended.

Modules

__rt

database

Traits to represent a database driver.

decode

Provides Decode for decoding values from the database.

encode

Provides Encode for encoding values for the database.

error

Types for working with errors produced by SQLx.

migrate

pool

Provides the connection pool for asynchronous SQLx connections.

prelude

Convenience re-export of common traits.

query

Types and traits for the query family of functions and macros.

query_builder

Runtime query-builder API.

sqlite

SQLite database driver.

sqlx_macros

testing

types

Conversions between Rust and SQL types.

Macros

migrate

Embeds migrations into the binary by expanding to a static instance of Migrator.

query

Statically checked SQL query with println!() style syntax.

query_as

A variant of [query!] which takes a path to an explicitly defined struct as the output type.

query_as_unchecked

A variant of [query_as!] which does not check the input or output types. This still does parse the query to ensure it’s syntactically and semantically valid for the current database.

query_file

A variant of [query!] where the SQL query is stored in a separate file.

query_file_as

Combines the syntaxes of [query_as!] and [query_file!].

query_file_as_unchecked

A variant of [query_file_as!] which does not check the input or output types. This still does parse the query to ensure it’s syntactically and semantically valid for the current database.

query_file_scalar

A variant of query_scalar! which takes a file path like [query_file!].

query_file_scalar_unchecked

A variant of query_file_scalar! which does not typecheck bind parameters and leaves the output type to inference. The query itself is still checked that it is syntactically and semantically valid for the database, that it only produces one column and that the number of bind parameters is correct.

query_file_unchecked

A variant of [query_file!] which does not check the input or output types. This still does parse the query to ensure it’s syntactically and semantically valid for the current database.

query_scalar

A variant of [query!] which expects a single column from the query and evaluates to an instance of QueryScalar.

query_scalar_unchecked

A variant of query_scalar! which does not typecheck bind parameters and leaves the output type to inference. The query itself is still checked that it is syntactically and semantically valid for the database, that it only produces one column and that the number of bind parameters is correct.

query_unchecked

A variant of [query!] which does not check the input or output types. This still does parse the query to ensure it’s syntactically and semantically valid for the current database.

Structs

Pool

An asynchronous pool of SQLx database connections.

QueryBuilder

A builder type for constructing queries at runtime.

RawSql

One or more raw SQL statements, separated by semicolons (;).

Sqlite

Sqlite database driver.

SqliteConnection

A connection to an open Sqlite database.

Transaction

An in-progress database transaction or savepoint.

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Error

Represents all the ways a method can fail within SQLx.

Traits

Acquire

Acquire connections or transactions from a database in a generic way.

Arguments

A tuple of arguments to be sent to the database.

Column

ColumnIndex

A type that can be used to index into a Row or Statement.

ConnectOptions

Connection

Represents a single database connection.

Database

A database driver.

Decode

A type that can be decoded from the database.

Encode

Encode a single value to be sent to the database.

Execute

A type that may be executed against a database connection.

Executor

A type that contains or can provide a database connection to use for executing queries against the database.

FromRow

A record that can be built from a row returned by the database.

IntoArguments

Row

Represents a single row from the database.

SqliteExecutor

An alias for Executor<'_, Database = Sqlite>.

Statement

An explicitly prepared statement.

Type

Indicates that a SQL type is supported for a database.

TypeInfo

Provides information about a SQL type for the database driver.

Value

An owned value from the database.

ValueRef

A reference to a single value from the database.

Functions

query

Execute a single SQL query as a prepared statement (transparently cached).

query_as

Execute a single SQL query as a prepared statement (transparently cached). Maps rows to Rust types using FromRow.

query_as_with

Execute a single SQL query, with the given arguments as a prepared statement (transparently cached). Maps rows to Rust types using FromRow.

query_scalar

Execute a single SQL query as a prepared statement (transparently cached) and extract the first column of each row.

query_scalar_with

Execute a SQL query as a prepared statement (transparently cached), with the given arguments, and extract the first column of each row.

query_with

Execute a SQL query as a prepared statement (transparently cached), with the given arguments.

raw_sql

Execute one or more statements as raw SQL, separated by semicolons (;).

test_block_on

Type Aliases

Result

A specialized Result type for SQLx.

SqlitePool

An alias for Pool, specialized for SQLite.

Attribute Macros

test

Derive Macros

Decode

Encode

FromRow

Type