hoprd_migration::sea_orm::sqlx
Trait Arguments

pub trait Arguments<'q>:
    Sized
    + Send
    + Default {
    type Database: Database;

    // Required methods
    fn reserve(&mut self, additional: usize, size: usize);
    fn add<T>(&mut self, value: T)
       where T: 'q + Send + Encode<'q, Self::Database> + Type<Self::Database>;

    // Provided method
    fn format_placeholder<W>(&self, writer: &mut W) -> Result<(), Error>
       where W: Write { ... }
}

A tuple of arguments to be sent to the database.

Required Associated Types

type Database: Database

Required Methods

fn reserve(&mut self, additional: usize, size: usize)

Reserves the capacity for at least additional more values (of size total bytes) to be added to the arguments without a reallocation.

fn add<T>(&mut self, value: T)
where T: 'q + Send + Encode<'q, Self::Database> + Type<Self::Database>,

Add the value to the end of the arguments.

Provided Methods

fn format_placeholder<W>(&self, writer: &mut W) -> Result<(), Error>
where W: Write,

Object Safety

This trait is not object safe.