hoprd_migration::sea_orm::sqlx
Trait Execute

pub trait Execute<'q, DB>: Sized + Send
where
    DB: Database,
{
    // Required methods
    fn sql(&self) -> &'q str;
    fn statement(&self) -> Option<&<DB as HasStatement<'q>>::Statement>;
    fn take_arguments(&mut self) -> Option<<DB as HasArguments<'q>>::Arguments>;
    fn persistent(&self) -> bool;
}

A type that may be executed against a database connection.

Implemented for the following:

&str

Query

Required Methods

fn sql(&self) -> &'q str

Gets the SQL that will be executed.

fn statement(&self) -> Option<&<DB as HasStatement<'q>>::Statement>

Gets the previously cached statement, if available.

fn take_arguments(&mut self) -> Option<<DB as HasArguments<'q>>::Arguments>

Returns the arguments to be bound against the query string.

Returning None for Arguments indicates to use a “simple” query protocol and to not prepare the query. Returning Some(Default::default()) is an empty arguments object that will be prepared (and cached) before execution.

fn persistent(&self) -> bool

Returns true if the statement should be cached.

Object Safety

This trait is not object safe.