hoprd_migration::sea_orm::sqlx::types
Struct Decimal

pub struct Decimal { /* private fields */ }

Decimal represents a 128 bit representation of a fixed-precision decimal number. The finite set of values of type Decimal are of the form m / 10e, where m is an integer such that -296 < m < 296, and e is an integer between 0 and 28 inclusive.

Implementations

impl Decimal

pub const MIN: Decimal = MIN

The smallest value that can be represented by this decimal type.

Examples

Basic usage:

assert_eq!(Decimal::MIN, dec!(-79_228_162_514_264_337_593_543_950_335));

pub const MAX: Decimal = MAX

The largest value that can be represented by this decimal type.

Examples

Basic usage:

assert_eq!(Decimal::MAX, dec!(79_228_162_514_264_337_593_543_950_335));

pub const ZERO: Decimal = ZERO

A constant representing 0.

Examples

Basic usage:

assert_eq!(Decimal::ZERO, dec!(0));

pub const ONE: Decimal = ONE

A constant representing 1.

Examples

Basic usage:

assert_eq!(Decimal::ONE, dec!(1));

pub const NEGATIVE_ONE: Decimal = NEGATIVE_ONE

A constant representing -1.

Examples

Basic usage:

assert_eq!(Decimal::NEGATIVE_ONE, dec!(-1));

pub const TWO: Decimal = TWO

A constant representing 2.

Examples

Basic usage:

assert_eq!(Decimal::TWO, dec!(2));

pub const TEN: Decimal = TEN

A constant representing 10.

Examples

Basic usage:

assert_eq!(Decimal::TEN, dec!(10));

pub const ONE_HUNDRED: Decimal = ONE_HUNDRED

A constant representing 100.

Examples

Basic usage:

assert_eq!(Decimal::ONE_HUNDRED, dec!(100));

pub const ONE_THOUSAND: Decimal = ONE_THOUSAND

A constant representing 1000.

Examples

Basic usage:

assert_eq!(Decimal::ONE_THOUSAND, dec!(1000));

pub fn new(num: i64, scale: u32) -> Decimal

Returns a Decimal with a 64 bit m representation and corresponding e scale.

Arguments

num - An i64 that represents the m portion of the decimal number

scale - A u32 representing the e portion of the decimal number.

Panics

This function panics if scale is > 28.

Example

let pi = Decimal::new(3141, 3);
assert_eq!(pi.to_string(), "3.141");

pub const fn try_new(num: i64, scale: u32) -> Result<Decimal, Error>

Checked version of Decimal::new. Will return Err instead of panicking at run-time.

Example

let max = Decimal::try_new(i64::MAX, u32::MAX);
assert!(max.is_err());

pub fn from_i128_with_scale(num: i128, scale: u32) -> Decimal

Creates a Decimal using a 128 bit signed m representation and corresponding e scale.

Arguments

num - An i128 that represents the m portion of the decimal number

scale - A u32 representing the e portion of the decimal number.

Panics

This function panics if scale is > 28 or if num exceeds the maximum supported 96 bits.

Example

let pi = Decimal::from_i128_with_scale(3141i128, 3);
assert_eq!(pi.to_string(), "3.141");

pub const fn try_from_i128_with_scale( num: i128, scale: u32, ) -> Result<Decimal, Error>

Checked version of Decimal::from_i128_with_scale. Will return Err instead of panicking at run-time.

Example

let max = Decimal::try_from_i128_with_scale(i128::MAX, u32::MAX);
assert!(max.is_err());

pub const fn from_parts( lo: u32, mid: u32, hi: u32, negative: bool, scale: u32, ) -> Decimal

Returns a Decimal using the instances constituent parts.

Arguments

lo - The low 32 bits of a 96-bit integer.

mid - The middle 32 bits of a 96-bit integer.

hi - The high 32 bits of a 96-bit integer.

negative - true to indicate a negative number.

scale - A power of 10 ranging from 0 to 28.

Caution: Undefined behavior

While a scale greater than 28 can be passed in, it will be automatically capped by this function at the maximum precision. The library opts towards this functionality as opposed to a panic to ensure that the function can be treated as constant. This may lead to undefined behavior in downstream applications and should be treated with caution.

Example

let pi = Decimal::from_parts(1102470952, 185874565, 1703060790, false, 28);
assert_eq!(pi.to_string(), "3.1415926535897932384626433832");

pub fn from_scientific(value: &str) -> Result<Decimal, Error>

Returns a Result which if successful contains the Decimal constitution of the scientific notation provided by value.

Arguments

value - The scientific notation of the Decimal.

Example

let value = Decimal::from_scientific("9.7e-7")?;
assert_eq!(value.to_string(), "0.00000097");

pub fn from_str_radix(str: &str, radix: u32) -> Result<Decimal, Error>

Converts a string slice in a given base to a decimal.

The string is expected to be an optional + sign followed by digits. Digits are a subset of these characters, depending on radix, and will return an error if outside the expected range:

0-9

a-z

A-Z

Examples

Basic usage:

assert_eq!(Decimal::from_str_radix("A", 16)?.to_string(), "10");

pub fn from_str_exact(str: &str) -> Result<Decimal, Error>

Parses a string slice into a decimal. If the value underflows and cannot be represented with the given scale then this will return an error.

Examples

Basic usage:

assert_eq!(Decimal::from_str_exact("0.001")?.to_string(), "0.001");
assert_eq!(Decimal::from_str_exact("0.00000_00000_00000_00000_00000_001")?.to_string(), "0.0000000000000000000000000001");
assert_eq!(Decimal::from_str_exact("0.00000_00000_00000_00000_00000_0001"), Err(Error::Underflow));

pub const fn scale(&self) -> u32

Returns the scale of the decimal number, otherwise known as e.

Example

let num = Decimal::new(1234, 3);
assert_eq!(num.scale(), 3u32);

pub const fn mantissa(&self) -> i128

Returns the mantissa of the decimal number.

Example

use rust_decimal_macros::dec;

let num = dec!(-1.2345678);
assert_eq!(num.mantissa(), -12345678i128);
assert_eq!(num.scale(), 7);

pub const fn is_zero(&self) -> bool

Returns true if this Decimal number is equivalent to zero.

Example

let num = Decimal::ZERO;
assert!(num.is_zero());

pub fn is_integer(&self) -> bool

Returns true if this Decimal number has zero fractional part (is equal to an integer)

Example

assert_eq!(dec!(5).is_integer(), true);
// Trailing zeros are also ignored
assert_eq!(dec!(5.0000).is_integer(), true);
// If there is a fractional part then it is not an integer
assert_eq!(dec!(5.1).is_integer(), false);

pub fn set_sign(&mut self, positive: bool)

๐Deprecated since 1.4.0: please use set_sign_positive instead

An optimized method for changing the sign of a decimal number.

Arguments

positive: true if the resulting decimal should be positive.

Example

let mut one = Decimal::ONE;
one.set_sign(false);
assert_eq!(one.to_string(), "-1");

pub fn set_sign_positive(&mut self, positive: bool)

An optimized method for changing the sign of a decimal number.

Arguments

positive: true if the resulting decimal should be positive.

Example

let mut one = Decimal::ONE;
one.set_sign_positive(false);
assert_eq!(one.to_string(), "-1");

pub fn set_sign_negative(&mut self, negative: bool)

An optimized method for changing the sign of a decimal number.

Arguments

negative: true if the resulting decimal should be negative.

Example

let mut one = Decimal::ONE;
one.set_sign_negative(true);
assert_eq!(one.to_string(), "-1");

pub fn set_scale(&mut self, scale: u32) -> Result<(), Error>

An optimized method for changing the scale of a decimal number.

Arguments

scale: the new scale of the number

Example

let mut one = Decimal::ONE;
one.set_scale(5)?;
assert_eq!(one.to_string(), "0.00001");

pub fn rescale(&mut self, scale: u32)

Modifies the Decimal towards the desired scale, attempting to do so without changing the underlying number itself.

Setting the scale to something less then the current Decimals scale will cause the newly created Decimal to perform rounding using the MidpointAwayFromZero strategy.

Scales greater than the maximum precision that can be represented by Decimal will be automatically rounded to either Decimal::MAX_PRECISION or the maximum precision that can be represented with the given mantissa.

Arguments

scale: The desired scale to use for the new Decimal number.

Example

use rust_decimal_macros::dec;

// Rescaling to a higher scale preserves the value
let mut number = dec!(1.123);
assert_eq!(number.scale(), 3);
number.rescale(6);
assert_eq!(number.to_string(), "1.123000");
assert_eq!(number.scale(), 6);

// Rescaling to a lower scale forces the number to be rounded
let mut number = dec!(1.45);
assert_eq!(number.scale(), 2);
number.rescale(1);
assert_eq!(number.to_string(), "1.5");
assert_eq!(number.scale(), 1);

// This function never fails. Consequently, if a scale is provided that is unable to be
// represented using the given mantissa, then the maximum possible scale is used.
let mut number = dec!(11.76470588235294);
assert_eq!(number.scale(), 14);
number.rescale(28);
// A scale of 28 cannot be represented given this mantissa, however it was able to represent
// a number with a scale of 27
assert_eq!(number.to_string(), "11.764705882352940000000000000");
assert_eq!(number.scale(), 27);

pub const fn serialize(&self) -> [u8; 16]

Returns a serialized version of the decimal number. The resulting byte array will have the following representation:

Bytes 1-4: flags

Bytes 5-8: lo portion of m

Bytes 9-12: mid portion of m

Bytes 13-16: high portion of m

pub fn deserialize(bytes: [u8; 16]) -> Decimal

Deserializes the given bytes into a decimal number. The deserialized byte representation must be 16 bytes and adhere to the following convention:

Bytes 1-4: flags

Bytes 5-8: lo portion of m

Bytes 9-12: mid portion of m

Bytes 13-16: high portion of m

pub fn is_negative(&self) -> bool

๐Deprecated since 0.6.3: please use is_sign_negative instead

Returns true if the decimal is negative.

pub fn is_positive(&self) -> bool

๐Deprecated since 0.6.3: please use is_sign_positive instead

Returns true if the decimal is positive.

pub const fn is_sign_negative(&self) -> bool

Returns true if the sign bit of the decimal is negative.

Example

assert_eq!(true, Decimal::new(-1, 0).is_sign_negative());
assert_eq!(false, Decimal::new(1, 0).is_sign_negative());

pub const fn is_sign_positive(&self) -> bool

Returns true if the sign bit of the decimal is positive.

Example

assert_eq!(false, Decimal::new(-1, 0).is_sign_positive());
assert_eq!(true, Decimal::new(1, 0).is_sign_positive());

pub const fn min_value() -> Decimal

๐Deprecated since 1.12.0: Use the associated constant Decimal::MIN

Returns the minimum possible number that Decimal can represent.

pub const fn max_value() -> Decimal

๐Deprecated since 1.12.0: Use the associated constant Decimal::MAX

Returns the maximum possible number that Decimal can represent.

pub fn trunc(&self) -> Decimal

Returns a new Decimal integral with no fractional portion. This is a true truncation whereby no rounding is performed.

Example

let pi = dec!(3.141);
assert_eq!(pi.trunc(), dec!(3));

// Negative numbers are similarly truncated without rounding
let neg = dec!(-1.98765);
assert_eq!(neg.trunc(), Decimal::NEGATIVE_ONE);

pub fn trunc_with_scale(&self, scale: u32) -> Decimal

Returns a new Decimal with the fractional portion delimited by scale. This is a true truncation whereby no rounding is performed.

Example

let pi = dec!(3.141592);
assert_eq!(pi.trunc_with_scale(2), dec!(3.14));

// Negative numbers are similarly truncated without rounding
let neg = dec!(-1.98765);
assert_eq!(neg.trunc_with_scale(1), dec!(-1.9));

pub fn fract(&self) -> Decimal

Returns a new Decimal representing the fractional portion of the number.

Example

let pi = Decimal::new(3141, 3);
let fract = Decimal::new(141, 3);
// note that it returns a decimal
assert_eq!(pi.fract(), fract);

pub fn abs(&self) -> Decimal

Computes the absolute value of self.

Example

let num = Decimal::new(-3141, 3);
assert_eq!(num.abs().to_string(), "3.141");

pub fn floor(&self) -> Decimal

Returns the largest integer less than or equal to a number.

Example

let num = Decimal::new(3641, 3);
assert_eq!(num.floor().to_string(), "3");

pub fn ceil(&self) -> Decimal

Returns the smallest integer greater than or equal to a number.

Example

let num = Decimal::new(3141, 3);
assert_eq!(num.ceil().to_string(), "4");
let num = Decimal::new(3, 0);
assert_eq!(num.ceil().to_string(), "3");

pub fn max(self, other: Decimal) -> Decimal

Returns the maximum of the two numbers.

let x = Decimal::new(1, 0);
let y = Decimal::new(2, 0);
assert_eq!(y, x.max(y));

pub fn min(self, other: Decimal) -> Decimal

Returns the minimum of the two numbers.

let x = Decimal::new(1, 0);
let y = Decimal::new(2, 0);
assert_eq!(x, x.min(y));

pub fn normalize(&self) -> Decimal

Strips any trailing zeroโs from a Decimal and converts -0 to 0.

Example

let number = Decimal::from_str("3.100")?;
assert_eq!(number.normalize().to_string(), "3.1");

pub fn normalize_assign(&mut self)

An in place version of normalize. Strips any trailing zeroโs from a Decimal and converts -0 to 0.

Example

let mut number = Decimal::from_str("3.100")?;
assert_eq!(number.to_string(), "3.100");
number.normalize_assign();
assert_eq!(number.to_string(), "3.1");

pub fn round(&self) -> Decimal

Returns a new Decimal number with no fractional portion (i.e. an integer). Rounding currently follows โBankers Roundingโ rules. e.g. 6.5 -> 6, 7.5 -> 8

Example

// Demonstrating bankers rounding...
let number_down = Decimal::new(65, 1);
let number_up   = Decimal::new(75, 1);
assert_eq!(number_down.round().to_string(), "6");
assert_eq!(number_up.round().to_string(), "8");

pub fn round_dp_with_strategy( &self, dp: u32, strategy: RoundingStrategy, ) -> Decimal

Returns a new Decimal number with the specified number of decimal points for fractional portion. Rounding is performed using the provided [RoundingStrategy]

Arguments

dp: the number of decimal points to round to.

strategy: the [RoundingStrategy] to use.

Example

let tax = dec!(3.4395);
assert_eq!(tax.round_dp_with_strategy(2, RoundingStrategy::MidpointAwayFromZero).to_string(), "3.44");

pub fn round_dp(&self, dp: u32) -> Decimal

Returns a new Decimal number with the specified number of decimal points for fractional portion. Rounding currently follows โBankers Roundingโ rules. e.g. 6.5 -> 6, 7.5 -> 8

Arguments

dp: the number of decimal points to round to.

Example

let pi = dec!(3.1415926535897932384626433832);
assert_eq!(pi.round_dp(2).to_string(), "3.14");

pub fn round_sf(&self, digits: u32) -> Option<Decimal>

Returns Some(Decimal) number rounded to the specified number of significant digits. If the resulting number is unable to be represented by the Decimal number then None will be returned. When the number of significant figures of the Decimal being rounded is greater than the requested number of significant digits then rounding will be performed using MidpointNearestEven strategy.

Arguments

digits: the number of significant digits to round to.

Remarks

A significant figure is determined using the following rules:

Non-zero digits are always significant.

Zeros between non-zero digits are always significant.

Leading zeros are never significant.

Trailing zeros are only significant if the number contains a decimal point.

Example

use rust_decimal_macros::dec;

let value = dec!(305.459);
assert_eq!(value.round_sf(0), Some(dec!(0)));
assert_eq!(value.round_sf(1), Some(dec!(300)));
assert_eq!(value.round_sf(2), Some(dec!(310)));
assert_eq!(value.round_sf(3), Some(dec!(305)));
assert_eq!(value.round_sf(4), Some(dec!(305.5)));
assert_eq!(value.round_sf(5), Some(dec!(305.46)));
assert_eq!(value.round_sf(6), Some(dec!(305.459)));
assert_eq!(value.round_sf(7), Some(dec!(305.4590)));
assert_eq!(Decimal::MAX.round_sf(1), None);

let value = dec!(0.012301);
assert_eq!(value.round_sf(3), Some(dec!(0.0123)));

pub fn round_sf_with_strategy( &self, digits: u32, strategy: RoundingStrategy, ) -> Option<Decimal>

Returns Some(Decimal) number rounded to the specified number of significant digits. If the resulting number is unable to be represented by the Decimal number then None will be returned. When the number of significant figures of the Decimal being rounded is greater than the requested number of significant digits then rounding will be performed using the provided [RoundingStrategy].

Arguments

digits: the number of significant digits to round to.

strategy: if required, the rounding strategy to use.

Remarks

A significant figure is determined using the following rules:

Non-zero digits are always significant.

Zeros between non-zero digits are always significant.

Leading zeros are never significant.

Trailing zeros are only significant if the number contains a decimal point.

Example

use rust_decimal_macros::dec;

let value = dec!(305.459);
assert_eq!(value.round_sf_with_strategy(0, RoundingStrategy::ToZero), Some(dec!(0)));
assert_eq!(value.round_sf_with_strategy(1, RoundingStrategy::ToZero), Some(dec!(300)));
assert_eq!(value.round_sf_with_strategy(2, RoundingStrategy::ToZero), Some(dec!(300)));
assert_eq!(value.round_sf_with_strategy(3, RoundingStrategy::ToZero), Some(dec!(305)));
assert_eq!(value.round_sf_with_strategy(4, RoundingStrategy::ToZero), Some(dec!(305.4)));
assert_eq!(value.round_sf_with_strategy(5, RoundingStrategy::ToZero), Some(dec!(305.45)));
assert_eq!(value.round_sf_with_strategy(6, RoundingStrategy::ToZero), Some(dec!(305.459)));
assert_eq!(value.round_sf_with_strategy(7, RoundingStrategy::ToZero), Some(dec!(305.4590)));
assert_eq!(Decimal::MAX.round_sf_with_strategy(1, RoundingStrategy::ToZero), Some(dec!(70000000000000000000000000000)));

let value = dec!(0.012301);
assert_eq!(value.round_sf_with_strategy(3, RoundingStrategy::AwayFromZero), Some(dec!(0.0124)));

pub const fn unpack(&self) -> UnpackedDecimal

Convert Decimal to an internal representation of the underlying struct. This is useful for debugging the internal state of the object.

Important Disclaimer

This is primarily intended for library maintainers. The internal representation of a Decimal is considered โunstableโ for public use.

Example

use rust_decimal_macros::dec;

let pi = dec!(3.1415926535897932384626433832);
assert_eq!(format!("{:?}", pi), "3.1415926535897932384626433832");
assert_eq!(format!("{:?}", pi.unpack()), "UnpackedDecimal { \
    negative: false, scale: 28, hi: 1703060790, mid: 185874565, lo: 1102470952 \
}");

pub fn from_f32_retain(n: f32) -> Option<Decimal>

Parses a 32-bit float into a Decimal number whilst retaining any non-guaranteed precision.

Typically when a float is parsed in Rust Decimal, any excess bits (after ~7.22 decimal points for f32 as per IEEE-754) are removed due to any digits following this are considered an approximation at best. This function bypasses this additional step and retains these excess bits.

Example

// Usually floats are parsed leveraging float guarantees. i.e. 0.1_f32 => 0.1
assert_eq!("0.1", Decimal::from_f32(0.1_f32).unwrap().to_string());

// Sometimes, we may want to represent the approximation exactly.
assert_eq!("0.100000001490116119384765625", Decimal::from_f32_retain(0.1_f32).unwrap().to_string());

pub fn from_f64_retain(n: f64) -> Option<Decimal>

Parses a 64-bit float into a Decimal number whilst retaining any non-guaranteed precision.

Typically when a float is parsed in Rust Decimal, any excess bits (after ~15.95 decimal points for f64 as per IEEE-754) are removed due to any digits following this are considered an approximation at best. This function bypasses this additional step and retains these excess bits.

Example

// Usually floats are parsed leveraging float guarantees. i.e. 0.1_f64 => 0.1
assert_eq!("0.1", Decimal::from_f64(0.1_f64).unwrap().to_string());

// Sometimes, we may want to represent the approximation exactly.
assert_eq!("0.1000000000000000055511151231", Decimal::from_f64_retain(0.1_f64).unwrap().to_string());

impl Decimal

pub fn checked_add(self, other: Decimal) -> Option<Decimal>

Checked addition. Computes self + other, returning None if overflow occurred.

pub fn saturating_add(self, other: Decimal) -> Decimal

Saturating addition. Computes self + other, saturating at the relevant upper or lower boundary.

pub fn checked_mul(self, other: Decimal) -> Option<Decimal>

Checked multiplication. Computes self * other, returning None if overflow occurred.

pub fn saturating_mul(self, other: Decimal) -> Decimal

Saturating multiplication. Computes self * other, saturating at the relevant upper or lower boundary.

pub fn checked_sub(self, other: Decimal) -> Option<Decimal>

Checked subtraction. Computes self - other, returning None if overflow occurred.

pub fn saturating_sub(self, other: Decimal) -> Decimal

Saturating subtraction. Computes self - other, saturating at the relevant upper or lower boundary.

pub fn checked_div(self, other: Decimal) -> Option<Decimal>

Checked division. Computes self / other, returning None if overflow occurred.

pub fn checked_rem(self, other: Decimal) -> Option<Decimal>

Checked remainder. Computes self % other, returning None if overflow occurred.