hoprd_migration::sea_orm
Struct DatabaseTransaction

pub struct DatabaseTransaction { /* private fields */ }

Defines a database transaction, whether it is an open transaction and the type of backend to use

Implementations

impl DatabaseTransaction

pub async fn commit(self) -> Result<(), DbErr>

Commit a transaction atomically

pub async fn rollback(self) -> Result<(), DbErr>

rolls back a transaction in case error are encountered during the operation