hoprd_migration::sea_orm
Trait SelectorTrait

pub trait SelectorTrait {
    type Item;

    // Required method
    fn from_raw_query_result(res: QueryResult) -> Result<Self::Item, DbErr>;
}

A Trait for any type that can perform SELECT queries

Required Associated Types

type Item

Required Methods

fn from_raw_query_result(res: QueryResult) -> Result<Self::Item, DbErr>

The method to perform a query on a Model

Object Safety

This trait is not object safe.